Purple Houses entrances and cave
I did the entrances to the purple houses including the stairs in both diractions. tho theres a bug that in house 3 and 4 it show the reflection of the player on the wall like in water.

also i put a little darker floor in the crystal cave entrance down to make it look more obvious.